elastic_search_service/pyproject.toml
[tool.poetry]
name = "msuic service"
version = "0.1.0"
description = ""
authors = ["karseny99 <arseny.karimov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.72.0"
grpcio-tools = "^1.72.0"
pydantic-settings = "^2.9.1"
dependency-injector = "^4.46.0"
asyncpg = "^0.30.0"
faststream = {extras = ["kafka"], version = "^0.5.40"}
setuptools = "^80.8.0"

elasticsearch = { version = "^8.0.0", extras = ["async"] }
uvloop = "^0.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
